+++ examples/parallel.py
@@ -18,7 +18,7 @@ import unified_planning as up
 from unified_planning.model import EndTiming, StartTiming
 
 from up_esb.bridge import Bridge
-from up_esb.executor import Executor
+from up_esb.execution.parallel_executor import Executor
 
 
 # TODO: Better example
